moving files to fit repository pattern
