+++ experimental/reporting-ui/src/main/react/jest/BUILD.bazel
@@ -4,9 +4,9 @@ js_library(
     name = "jest",
     srcs = ["jest_base.config.js"],
     visibility = [
-        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
     ],
     deps = [
-        "//experimental/reporting-ui/src/test/react/jest/mocks",
+        "//experimental/reporting-ui/src/main/react/jest/mocks",
     ],
 )
