+++ experimental/reporting-ui/src/main/react/jest/mocks/BUILD.bazel
@@ -4,6 +4,6 @@ js_library(
     name = "mocks",
     srcs = ["styleMock.ts"],
     visibility = [
-        "//experimental/reporting-ui/src/test/react/jest:__subpackages__",
+        "//experimental/reporting-ui/src/main/react/jest:__subpackages__",
     ],
 )
