+++ experimental/reporting-ui/src/main/react/jest/mocks/styleMock.ts
