+++ experimental/reporting-ui/src/main/react/reporting-ui/client/reporting/BUILD.bazel
@@ -3,7 +3,6 @@ load("@aspect_rules_js//js:defs.bzl", "js_library")
 package(
     default_visibility = [
         "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
-        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
     ],
 )
