+++ experimental/reporting-ui/src/main/react/reporting-ui/component/loader/BUILD.bazel
@@ -3,14 +3,12 @@ load("@aspect_rules_js//js:defs.bzl", "js_library")
 package(
     default_visibility = [
         "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
-        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
     ],
 )
 
 js_library(
-    name = "general_component",
+    name = "loader",
     srcs = [
-        "error.tsx",
         "loader.css",
         "loader.tsx",
     ],
