+++ experimental/reporting-ui/src/main/react/reporting-ui/model/report_list/report_list_model.ts
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import appConfig from '../../../client/initialize';
+import appConfig from '../../client/initialize';
 
 export const ReportListModel = () => {
   async function loadReports() {
