+++ experimental/reporting-ui/src/main/react/reporting-ui/route.tsx
@@ -12,7 +12,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import { ReportListView } from "./view/report/list/report_list_view";
+import { ReportListView } from "./view/report_list/report_list_view";
+
 
 type route = {
     path: string,
