--- experimental/reporting-ui/src/test/react/reporting-ui/BUILD.bazel
@@ -1,38 +1,0 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-load("@aspect_rules_jest//jest:defs.bzl", "jest_test")
-
-js_library(
-    name = "jest_config",
-    srcs = ["jest.config.js"],
-)
-
-js_library(
-    name = "test_node_deps",
-    srcs = [
-        "//:node_modules/@testing-library/dom",
-        "//:node_modules/@testing-library/jest-dom",
-        "//:node_modules/@testing-library/react",
-        "//:node_modules/@testing-library/user-event",
-        "//:node_modules/@types/jest",
-        "//:node_modules/@types/testing-library__jest-dom",
-        "//:node_modules/jest",
-        "//:node_modules/jest-environment-jsdom",
-        "//:node_modules/jest-junit",
-        "//:node_modules/jest-transform-stub",
-        "//:node_modules/ts-jest",
-    ],
-)
-
-jest_test(
-    name = "test",
-    config = ":jest_config",
-    data = [
-        ":test_node_deps",
-        "//:tsconfig",
-        "//experimental/reporting-ui/src/main/react/reporting-ui:app",
-        "//experimental/reporting-ui/src/main/react/reporting-ui:node_deps",
-        "//experimental/reporting-ui/src/test/react/jest",
-        "//experimental/reporting-ui/src/test/react/reporting-ui/view/report/list",
-    ],
-    node_modules = "//:node_modules",
-)
